+++ lib/src/presentation/output_print_generator/table_builder.dart
@@ -1,13 +1,14 @@
 import 'package:pull_request_coverage/src/extensions/string.dart';
 
-class CliTableBuilder {
+class TableBuilder {
+  static const header = ["Report", "Current value", "Threshold", "Result"];
+  
   static const headerDivider = "-";
   static const columnDivider = "|";
+  
+  int get columnsLength => header.length;
 
-  final int columnsLength;
-  final List<String>? header;
-
-  CliTableBuilder({required this.columnsLength, this.header});
+  TableBuilder();
 
   final List<List<String>> table = [];
 
@@ -24,9 +25,14 @@ class CliTableBuilder {
     return "$filler${filler * (remaining ~/ 2)}$content${filler * (remaining ~/ 2)}$suffix$filler";
   }
 
-  String build() {
+  String build(bool surroundedByDivider) {
     final stringBuffer = StringBuffer();
     final columnSize = List.generate(columnsLength, (index) => 0);
+
+    void writeDividerOnBorder() {
+      if (surroundedByDivider) stringBuffer.write(columnDivider);
+    }
+
     for (var columnIndex = 0; columnIndex < columnsLength; columnIndex++) {
       for (var lineIndex = 0; lineIndex < table.length; lineIndex++) {
         if (table[lineIndex][columnIndex].getLengthWithNoColor() > columnSize[columnIndex]) {
@@ -34,33 +40,37 @@ class CliTableBuilder {
         }
       }
     }
-    final header = this.header;
-    if (header != null) {
-      for (var i = 0; i < header.length; i++) {
-        if (header[i].length > columnSize[i]) {
-          columnSize[i] = header[i].length;
-        }
-        stringBuffer.write(_createContent(header[i], columnSize[i], " "));
-        if (i != header.length - 1) {
-          stringBuffer.write(columnDivider);
-        }
+    writeDividerOnBorder();
+    for (var i = 0; i < header.length; i++) {
+      if (header[i].length > columnSize[i]) {
+        columnSize[i] = header[i].length;
       }
-      stringBuffer.writeln();
-      for (var i = 0; i < header.length; i++) {
-        stringBuffer.write(_createContent("", columnSize[i], "-"));
-        if (i != header.length - 1) {
-          stringBuffer.write(columnDivider);
-        }
+      stringBuffer.write(_createContent(header[i], columnSize[i], " "));
+      if (i != header.length - 1) {
+        stringBuffer.write(columnDivider);
       }
     }
+    writeDividerOnBorder();
+    stringBuffer.writeln();
+    writeDividerOnBorder();
+    for (var i = 0; i < header.length; i++) {
+      stringBuffer.write(_createContent("", columnSize[i], "-"));
+      if (i != header.length - 1) {
+        stringBuffer.write(columnDivider);
+      }
+    }
+    writeDividerOnBorder();
+
     for (var lineIndex = 0; lineIndex < table.length; lineIndex++) {
       stringBuffer.writeln();
+      writeDividerOnBorder();
       for (var columnIndex = 0; columnIndex < columnsLength; columnIndex++) {
         stringBuffer.write(_createContent(table[lineIndex][columnIndex], columnSize[columnIndex], " "));
         if (columnIndex < columnsLength - 1) {
           stringBuffer.write(columnDivider);
         }
       }
+      writeDividerOnBorder();
     }
     return stringBuffer.toString();
   }